+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - new version check
 - code style enforced with eslint
+- support for environmental variables and .env files
 
 ## [1.1.1] - 2018-03-01
 ### Changed